refactor(server-config): split into multiple services
