+++ src/features/init/services/init-service.ts
@@ -18,7 +18,7 @@ import { GithubConfigService } from "../../github/services/github-config-service
 import { ChalkService } from "../../logger/services/chalk-service";
 import { LoggerConfigMutatorService } from "../../logger/services/config/logger-config-mutator-service";
 import { LoggerService } from "../../logger/services/logger-service";
-import { ServerConfigService } from "../../server/services/server-config-service";
+import { ServerConfigService } from "../../server/services/config/server-config-service";
 import { ServerService } from "../../server/services/server-service";
 
 export class InitService {
