--- src/features/server/constants/server-config.ts
@@ -1,5 +1,0 @@
-import { IServerConfig } from "../interfaces/server-config";
-
-export const SERVER_CONFIG: IServerConfig = {
-  port: 3001,
-};
