+++ src/features/server/enums/server-config-value-name.enum.spec.ts
@@ -1,0 +1,9 @@
+import { ServerConfigValueNameEnum } from "./server-config-value-name.enum";
+
+describe(`ServerConfigValueNameEnum`, (): void => {
+  it(`should have a member "PORT"`, (): void => {
+    expect.assertions(1);
+
+    expect(ServerConfigValueNameEnum.PORT).toStrictEqual(`port`);
+  });
+});
