+++ src/features/server/enums/server-config-value-name.enum.ts
@@ -1,0 +1,3 @@
+export enum ServerConfigValueNameEnum {
+  PORT = `port`,
+}
